requirements.txt
streamlit
pandas
numpy
scikit-learn
datetime
matplotlib
seaborn
loguru
requests
plotly
python-dotenv
fastapi
psycopg2-binary
sqlalchemy
duckdb
duckdb-engine
uvicorn